requirements.txt
absl-py==2.3.1
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.1.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cssselect==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.0
fastapi==0.116.2
ffmpy==0.6.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
gast==0.6.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==3.50.2
gradio_client==0.6.1
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
kiwisolver==1.4.7
kubernetes==33.1.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.29
libclang==18.1.1
lxml==6.0.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.5.0
networkx==3.2.1
nibabel==5.3.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pillow==10.4.0
posthog==5.4.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pubmed==0.0.2.post0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.4
pypdf==6.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyquery==2.0.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
ruff==0.13.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
semantic-version==2.10.0
sentence-transformers==5.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.1
tomli==2.2.1
tomlkit==0.12.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
typer==0.17.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
